refactor: file structure and tuning
