+++ src/utils/types.ts
@@ -1,0 +1,4 @@
+export type DividerResult<
+  T extends string | string[],
+  F extends boolean = false,
+> = T extends string ? string[] : F extends true ? string[] : string[][];
