+++ tsconfig.json
@@ -7,7 +7,10 @@
     "target": "ES6",
     "declarationDir": "dist",
     "strict": true,
-    "declaration": true
+    "declaration": true,
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist", "tests"]
